improvement: narwhals perf and arg name changes (#4994)
Perf: This improves some query logic for narwhals for better batching of
expressions.
Migration: This migrates some name changes that are deprecated, but have
been out for a while
Wider support: in some cases, we upgrade narwhals (`nw.v1` to `nw_main`)
for better support for newer features like writing to csv. This code can
be cleaned up with `nw.v2` is put up.