fix: Allow @prop.setter in class definitions (#5004)
## 📝 Summary

fixes #4993

`@prop` capture fix is a visitor change, since variables defined in
class scope are not bound to be "unbounded" if they're set in the class
body. Filtering is done in the loop since this is order dependent.

I think there might be some pretty hairy `global` edge cases I haven't
taken the time to think through- but we can reserve it for a
"correctness" sprint

@akshayka