feat: Add unique transform to `mo.ui.dataframe` (#4970)
## 📝 Summary

Closes #3738.

## 🔍 Description of Changes

### Frontend

- Add icon to `mo.ui.dataframe` transform panel.
- Define `UniqueTransformSchema` with four options for `keep` parameter
(first, last, any, none) in
`frontend/src/plugins/impl/data-frames/schema.ts`.
  - The only odd one out is `any`, which is only for Polars dataframes.

### Backend

- Create handler and types for `UniqueTransform` in
`marimo/_plugins/ui/_impl/dataframes/transforms`.
- Implement unique transform for pandas, polars, and ibis dataframes.
- Write Python statement for each dataframe type in `print_code.py`.

## Driveby Changes

- Removed `frontend/src/plugins/impl/data-frames/python` as it appears
to be unused, but please let me know if I'm mistaken.
- Fix Pylint error about using `if` instead of `elif` for all `if`
blocks that return in `dataframes/apply.py`.
- I disabled Pylint somewhere along the way but left this one in anyway.
Happy to revert.

## Tests

- Wrote small unit tests for unique transform handler using all 4 `keep`
options.
- Updated snapshots.
- Caveat:
- Since the result of a `unique` transform appears to be
non-deterministic for polars and ibis dataframes, it seems that I
couldn't easily extend the tests for `print_code.py`. However, I did
include a notebook in the smoke tests in which I manually verified that
the python print code can be run and appears to return the same number
of rows as the plugin.

## Notes

- I can't figure out why the frontend build job is failing. All frontend
checks and tests are passing in my env, and `cd frontend; pnpm install
--no-frozen-lockfile` makes no changes to the lockfile.
- The unique transform result appears to be non-deterministic for polars
and ibis. There is a `maintain_order` argument that we could include for
polars dataframes, but not sure if there is one for ibis.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

Not sure who to tag here so I'll put @mscolnick.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>