fix: github lsp build script, language server path
The build script was calling mkdir on a directory that already
existed, and failing silently.

This change fixes that and also updates the script to terminate on
error so we don't get any more silent failures.

Additionally, this updates the path to the language server to not
include dist -- when generating with pnpm 10.11 on macOS, at least, no
such directory is present.