chore: cleanup config passing to HTML from server (#5120)
This is a start of removing some of the complexity taken on by passing
config via HTML dom elements. This also paves the way to clean up some
global state that has leaked over time.

There are some DOM elements still left around, but will handle those in
a followup