improvement: notebook schema serialization (#5131)
The first thing that should be reviewed is:
`marimo/_schemas/generated/notebook.yaml`.

This adds an **internal** serialization format that is not Python. This
is so we can quickly hydrate the UI in WebAssembly, without waiting for
pyodide + marimo to download in order to parse the notebook file. This
also has an added benefit of simplifying and giving a stronger schema
for the static export.

Still needs thorough testing, migrating the static export to this, and
some cleanup, but want to get the schema reviewed first