Add mechanism for injecting custom file stores (#5161)
## 📝 Summary

There is currently a fixed order of precedence of checking different
FileStores as the source for loading and saving a notebook. This works
exactly right for current deployments, and even has hardcoded references
to the online playground, but doesn't allow custom deployments to have
their own sources of data.

## 🔍 Description of Changes

Adds a `file_stores` property to the app config element which takes
global class names that are pushed onto the `notebookFileStore` via a
new method `CompositeFileStore.insert`.

This allows someone to write code like the following to load notebooks
from [PouchDB](https://pouchdb.com/)

```html
<marimo-app-config data-config='{"file_stores": ["PouchDBFileStore"]}' hidden></marimo-app-config>
```

```javascript
        window.PouchDBFileStore = class PouchDBFileStore {
            constructor() { ... }
            async saveFile(contents) { ... }
            async readFile() { ... }
        };
```

I'm open to suggestions for better ways to do this and not very attached
to this code

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@akshayka OR @mscolnick

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>