fix: (outlines)  Headings with the same name are highlighted multiple… (#5176)
Headings with the same name are highlighted multiple time in the table
of contents

fixes #2087

## 📝 Summary

fixes an issue related to the currently active header if there are
multiple headings of the same type with the same content.

fixes #2087

## 🔍 Description of Changes

similar to how it was already done for navigating this PR keeps track of
the occurrence of an id-header pair and highlights only the correct
occurrence

## 📋 Checklist

- [(/) ] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [(/) ] I have run the code and verified that it works as expected.

## 📜 Reviewers

@akshayka OR @mscolnick

---------

Co-authored-by: johannes <jscheider@mailbox.org>