feat: Allow selecting optional "extras" for missing packages (#5260)
Closes #5236 

I'm fairly pleased with the UI for this, but parsing to and from package
specifiers is a bit of a challenge. I wonder if having a richer datatype
we send back and forth from the front end could be more robust



https://github.com/user-attachments/assets/cf93040f-828b-4d93-a161-34e0f51ba56e

Couple of design considerations:

- ~It's common for some projects to have `test`/`testing-`/`dev`
optional dependencies. Should we hide those by default?~
- ~The multiselect menu is really custom and not very accessible. I
could try to find an unstyled UI primitive (e.g., Radix UI) or just use
checkboxes to make it better. I like the look but key navigation doesn't
work and I think that's not great.~ I reused our Checkbox components
now. The styling is more consistent with our design system, and with
native checkboxes keyboard navigation works better.