Extend `view-as` URI parameter to handle app view in edit mode (#5327)
## 📝 Summary

Extends the `view-as` URI parameter to acommodate `view-as=present`.
This causes the notebook to be shown in the `present` or "app" view by
default when served in edit mode.

This is useful because it gives a clean UI by default while still
allowing the user to drop into the `edit` mode if they so desire.

## 🔍 Description of Changes

When the user runs `marimo edit notebook.py` and navigates to
`http://localhost:2718?view-as=present`, the notebook will start in app
view mode (aka present mode) instead of edit mode.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable)
  - This is a small change
- [x] I have added tests for the changes made
  - This change is well covered by the existing URI parameter tests
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

@akshayka OR @mscolnick

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>