pref: async path (#5363)
This adds an `AsyncPath` class that implements `Path` but puts each
operation in a thread (non-blocking the event loop) and makes it async.
This is useful when the FS is mirror to a cloud bucket which operations
can take longer (200-400ms) due to network latency.

This uses `asyncio.to_thread` (introduced in 3.9+).

---------

Co-authored-by: dylan <dylan@marimo.io>