Fix matplotlib plots not rendering in HTML export (#5377)
Regression in #5370 seem to be related to two separate things:

1. Cell outputs appear to be in topological order instead of document
order
2. Rich media console outputs aren't rendered in the HTML export

The latter seems to be related to our recent schema/serialization work
where matplotlib plots get stuck in console outputs instead of being
promoted to main outputs for frontend rendering.

I think I'm close to a fix but I haven't been able to run a local build
to verify (output assets point to a CDN with invalid hashes), so marking
this as draft.

Fixes #5370

---------

Co-authored-by: Akshay Agrawal <akshay@marimo.io>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>