logging: downgrade warning to debug for session cache serialization (#5423)
Demote log warning when serializing the session view to a log info.
* The session cache writer does not seem to know the notebook-order of
cells. This doesn't seem to be a problem in practice (can run a notebook
and hydrate from a cached session view object just fine). But it's odd
that not having the notebook order does pose a problem when hydrating an
exported HTML object.

@akshayka

---------

Co-authored-by: Akshay Agrawal <akshay@marimo.io>