dev: Use islands as local asset server (#5436)
## 📝 Summary

quarto `0.14.7` is broken.

I'm having trouble confirming this is the issue by replicating it
locally, since I can't replicate the breakage. My dev served assets seem
to work in general. But `being islands` != `being wasm` as is, and I
think some of the logic around sockets gets lost (0.14.6 works, so this
may be the offending PR: #5403)

@akshayka

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>