Remove "vega" as top-level dependency
All runtime uses of "vega" were replaced with "vega-loader" in #5449,
and now the only remaining usage is for types.

This change removes "vega" entirely as a direct dependency and switches
to extracting necessary types from "vega-lite" (which we are using
directly and extensively). It should also simplify future updates, since
we no longer need to keep "vega" and "vega-lite" in sync: all types are
now derived from what "vega-lite" exports. Should also prevent
accidental re-imports from "vega" in the future.