feat: configuration to disable sharing actions (#5432)
## 📝 Summary

This adds an option to disable (hide) all the sharing functionality,
e.g., "Share > Publish HTML to web". The issue is tracked in #5323


![image](https://github.com/user-attachments/assets/b93fda55-ace7-43f0-b5a2-6e335f01eed9)

![image](https://github.com/user-attachments/assets/8714a450-9788-408b-9dc1-bddba3b84710)

## 🔍 Description of Changes

Introduces a new configuration setting "sharing", which is enabled by
default. When it is disabled all the sharing related actions are hidden
in the UI.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

@mscolnick

---------

Co-authored-by: Kyrre-Wahl-Kongsgard <kyrre.wahl.kongsgard@dnb.no>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>