Add destructive cell deletion confirmation (#5664)
Fixes #5658

This change adds a keymap configuration (`destructive_delete`, default
**True**) to enable controlled, destructive deletion of cells via
keyboard shortcut. Video:

- First deletion is "cheap" (`d = 10`), and automatically deleted
- Section deletion is "expensive", and adds confirmation step

<img width=500
src="https://github.com/user-attachments/assets/ddbe0836-c5ef-4b6c-b531-ac4427a27461"/>


The idea is to allow deletion of cells, but to have a confirmation step
(i.e., "are you sure you want to do that?") when we consider a deletion
_could_ be "expensive". The definition of expensive is up for question
but right now we can look at:

- Whether a cells execution time is above some threshold (e.g., 2s)
- Whether a cell declares variables used by other cells

If neither of those conditions are met, we automatically delete the
cell.

What I hope this unlocks is the ability to readily delete cells for the
common editing use case, and give pause for the moments where a deletion
might be costly or unintentional. My preference is to have opt-out
unless the user explicitly prefers stricter safeguards.

---------

Co-authored-by: Myles Scolnick <myles@marimo.io>