chore/fix: parse correct message for anywidget (#5741)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->
anywidgets were logging a lot of errors when I tested with Quak &
mosaic. Although nothing was broken with quak.
```typescript
Failed to parse message ZodError: [
  {
    "code": "invalid_union_discriminator",
    "options": [
      "open",
      "update",
      "custom",
      "echo_update",
      "close"
    ],
    "path": [
      "method"
    ],
    "message": "Invalid discriminator value. Expected 'open' | 'update' | 'custom' | 'echo_update' | 'close'"
  }
]
```

It's because we are trying to parse the message's content rather than
the entire message. The following is the content, which does not have
the 'method' field.
```typescript
Message
{
    "type": "arrow",
    "uuid": "160ab694-ef1f-475a-b108-adc3fee81267"
}
```

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.