Add text wrapping toggle for console output (#5739)
Fixes #4968

Adds a button to toggle text wrapping in console outputs, making it
easier to read and analyze text without horizontal scrolling.

The toggle appears on hover in the top-right corner of console outputs.
Its state is currently persisted in localStorage, making it consistent
across cells (and sessions). Alternatively, we could:

- Add a marimo-wide configuration option, or
- Not persist anything and have each output control it's wrapping state



https://github.com/user-attachments/assets/96aa6ccd-96ed-4ece-8d4d-2d9f570946ea