Don't persist minimap state open/closed state between page reloads (#5757)
Because it hides parts of the UI when open, persisting minimap
open/closed state to localStorage is a bit awkward. Sometimes you close
a notebook with a complex minimap, then open a fresh one with no cells
(empty minimap). I think a better UX is to just have global state per
page reload.