fix: add suspense for vega charts (#5770)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->
There was an issue with the notebook flickering due to loading vega.
This adds a suspense and minor refactor on the frontend.

This also applies some auto-options for vegafusion charts (like making
the width full).

- I wanted to fix the background colour bug (which only happens with
vegafusion), but vegafusion by default sets it to white so I'm not keen
to overwrite it (we may clash with user specified background). We could
'smartly' check if theme == dark, then set to black, but it's a bit
hacky and imo we are interfering with defaults.

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.