fix: dynamic SQL keyword completion based on dialect (#5775)
This makes the keyword completion dynamic, using the dialect for the
selected connection. Previously it was just always using the standard
SQL dialect.

This also fixes caching - it was always "missing" the cache (it always
created a new object which gets passed to a Map and check for object
equality)
This also adds a few tests.