tests: cache invaliadation for external decorators (#5758)
## 📝 Summary

External functions are assumed to be pure. If an external function wraps
a notebook function, then our test was unable to pick up the notebook
change. This PR adds test to catch this behavior, and adds an
appropriate fix (explicitly checking to see if the external function is
used as a wrapper)

The PR also catches the general case of pickle not correctly capturing
the transitive references of unhashable ContextExecution objects. This
is a detectable edgecase with poorly defined behavior (by virtue of
pickle). Future work should include adding a lint rule encouraging users
away from this condition.