Bypass delete confirmation for empty cells (#5804)
An empty cell indicates the users intent to delete the cell, regardless
of the runtime state or variables it declares. This change marks empty
cells as "cheap" in our pending delete service, as well as only tracks
relationships to non-empty cells.

The latter allows us to delete other cells in bulk that are linked to
the "expensive" empty cell. For example:

```diff
# cell-1
import time

# cell-2
-- time.sleep(3)
++
```

we can delete cell-1 and cell-2 without confirmation.