Make reusable cell link only clickable on text, not entire block
Now only the "reusable" text and the icons are clickable links.
Previously, the entire output area block was clickable as a hyperlink
when a cell contained a reusable definition. This made it easy to
accidentally navigate away when clicking anywhere in that area, even
though most of the block is not visually indicated as a link.